+++ bin/po/po_translate.sh
@@ -23,7 +23,8 @@ locate_python
 if [[ ! -e $PYTHON_SITEDIR/polib.py &&
 	  ! -e $PYTHON_USER_SITEDIR/polib.py ]];
 then
-    $PYTHON -m pip install polib transformers translate torch sentencepiece sacremoses
+    $PYTHON -m pip install polib translate
+    #$PYTHON -m pip install polib transformers translate torch sentencepiece sacremoses
 fi
 
 $PYTHON bin/python/po_translate.py $*
@@ -31,4 +32,4 @@ $PYTHON bin/python/po_translate.py $*
 #
 # Remove all the dependencies we installed
 #
-$PYTHON -m pip uninstall -r requirements_useless.txt -y
+#$PYTHON -m pip uninstall -r requirements_useless.txt -y