+++ src/docs/HISTORY.md
@@ -1,3 +1,13 @@
+v1.3.7
+======
+
+- Fixed Ghost Previous and Ghost Next not being in sync on the Secondary 
+  viewport.
+- Made stepping with Next Annotation and Previous Annotation loop once they 
+  reach the final and first annotation respectively.
+- Fixed opening Secondary Window not displaying the video when the video was
+  stopped.
+  
 v1.3.6
 ======